fix : 비행검사 목록 삭제 전 confirm 후 삭제

+++ src/components/flight/FlightItem.tsx
@@ -66,8 +66,11 @@ const FlightItem = ({ testName, testType, testDate, id, refetch }: FlightList & 
 
     const DeleteFlightItem = (e: any) => {
         e.stopPropagation();
-        deleteFlightList(id).then(() => refetch())
         console.log('delete');
+
+        if(window.confirm("진짜로 비행점검 삭제?")) {            
+            deleteFlightList(id).then(() => refetch())
+        }
     }
 
     return (
